requirements.txt
Flask==2.0.3
pycryptodome==3.15.0
Flask-SocketIO==5.2.0
eventlet==0.33.0